moved the mailNodeMailer.ts to lib from routes
